speed_monitor_dashboard/requirements.txt
streamlit
pandas
numpy
folium
streamlit-folium
plotly
requests
Pillow
